Fix a compilation warning
git-svn-id: http://root.cern.ch/svn/root/trunk@31366 27541ba8-7e3a-0410-8455-c3a389f83636

+++ graf2d/qt/src/TQtPadFont.cxx
@@ -308,7 +308,7 @@ void   TQtPadFont::SetTextMagnify(Float_t  mgn)
    //
     Int_t tsize = (Int_t)(fTextSize+0.5);
     if (TMath::Abs(mgn-1) >0.05)  {
-       int pxSize = mgn*FontMagicFactor(tsize);
+       int pxSize = int(mgn*FontMagicFactor(tsize));
        if(pxSize<=0) pxSize=1;
        this->setPixelSize(pxSize);
     }
